requirements.txt
rembg
gradio==3.50.2
Pillow==10.2.0
numpy==1.26.4 
onnxruntime